classLoader uprava

+++ JVM/runtime/OperandStack.cpp
@@ -3,7 +3,7 @@
 
 OperandStack::OperandStack(): OperandStack(OperandStack::DEFAULT_STACK_SIZE) {}
 
-OperandStack::OperandStack(int size): Array(size)
+OperandStack::OperandStack(size_t size): Array(size)
 {
 }
